requirements.txt
flask
pystray
Pillow
pynput
psutil
# MacOS
pyobjc-framework-Cocoa; sys_platform == 'darwin'
pyobjc-framework-Quartz; sys_platform == 'darwin'
# Windows
pywin32; sys_platform == 'win32'
uiautomation; sys_platform == 'win32'
win10toast; sys_platform == 'win32'
google-generativeai
